+++ geo/CHANGES.md
@@ -13,6 +13,7 @@
 - Bump geo MSRV to 1.82
 - Simplify test rustc and libproj version specification in CI
 - Performance: Avoid running through entire iterator to reach last element in `outlier_detection` when calculating LRD and LOF
+- Add `Bearing` and `Destination` trait implementations for `Euclidean`
 
 ## 0.30.0 - 2025-03-24